feat: renamed Group, DM and Home feed
