+++ LMChatCore_iOS/LMChatCore_iOS/Source/Helper/LMCoreComponents.swift
@@ -14,8 +14,8 @@ public struct LMCoreComponents {
     public static var shared = Self()
     
     // MARK: HomeFeed Screen
-    public var groupChatFeedScreen: LMChatGroupFeedViewController.Type = LMChatGroupFeedViewController.self
-    public var chatFeedScreen: LMChatFeedViewController.Type = LMChatFeedViewController.self
+    public var groupChatFeedScreen: LMCommunityChatViewController.Type = LMCommunityChatViewController.self
+    public var chatFeedScreen: LMCommunityHybridChatViewController.Type = LMCommunityHybridChatViewController.self
     
     public var exploreChatroomListScreen: LMExploreChatroomListView.Type = LMExploreChatroomListView.self
     public var exploreChatroomScreen: LMExploreChatroomViewController.Type = LMExploreChatroomViewController.self
@@ -40,7 +40,7 @@ public struct LMCoreComponents {
     public var searchConversationListScreen: LMChatSearchConversationListViewController.Type = LMChatSearchConversationListViewController.self
     
     //MARK: DM Screen
-    public var dmChatFeedScreen: LMChatDMFeedViewController.Type = LMChatDMFeedViewController.self
+    public var dmChatFeedScreen: LMNetworkingChatViewController.Type = LMNetworkingChatViewController.self
     public var dmMemberListScreen: LMChatMemberListViewController.Type = LMChatMemberListViewController.self
     
     //MARK: Poll
