+++ LMChatCore_iOS/LMChatCore_iOS/Source/Module/NavigationAction/NavigationActions.swift
@@ -45,7 +45,7 @@ class NavigationScreen: NavigationScreenProtocol {
     func perform(_ action: NavigationActions, from source: UIViewController, params: Any?) {
         switch action {
         case .homeFeed:
-            guard let homefeedvc = try? LMChatGroupFeedViewModel.createModule() else { return }
+            guard let homefeedvc = try? LMCommunityChatViewModel.createModule() else { return }
             source.navigationController?.pushViewController(homefeedvc, animated: true)
         case .chatroom(let chatroomId, let conversationId):
             guard let chatroom = try? LMChatMessageListViewModel.createModule(withChatroomId: chatroomId, conversationId: conversationId) else { return }
