+++ LikemindsChatSample/LikemindsChatSample/ViewController.swift
@@ -150,7 +150,7 @@ class HomeViewController: UIViewController {
 
     @objc private func chatButtonTapped() {
         do {
-            let chatFeedVC = try LMChatFeedViewModel.createModule()
+            let chatFeedVC = try LMCommunityHybridChatViewModel.createModule()
             navigationController?.pushViewController(chatFeedVC, animated: true)
         } catch {
             let alert = UIAlertController(
